requirements.txt
chromadb>=0.4.24
transformers>=4.41.0
torch>=2.3.0
accelerate>=0.30.0
safetensors>=0.4.3
tqdm>=4.66.4
python-dotenv>=1.0.1
torch==2.3.0
torchvision==0.18.0
torchaudio==2.3.0
sentence-transformers==2.7.0
umap-learn==0.5.5
transformers==4.40.0
accelerate>=0.29.0
bitsandbytes>=0.43.0
tqdm==4.66.2
numpy==1.26.4
scikit-learn==1.4.2
pandas==2.2.1
torch
transformers
accelerate
faiss-gpu
gitpython
numpy
requests
tqdm
huggingface-hub
bitsandbytes
flash-attn